+++ tests/unit/data/commentsService.test.ts
@@ -10,7 +10,7 @@ describe('Comments Service', () => {
   const { schemaId } = commentData;
   const documentId = '2e9fff9d90135a2a9a718e2f';
   const commentId = commentData.id;
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
   let sdk: Client<ParamsOauth2>;
 
   beforeAll(async () => {