+++ tests/unit/data/schemasService.test.ts
@@ -16,7 +16,7 @@ import { createPagedResponse } from '../../__helpers__/utils';
 describe('Schemas Service', () => {
   const schemasListResponse = createPagedResponse(schemaData);
   const schemaId = newSchemaCreated.id;
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
   let sdk: Client<ParamsOauth2>;
 
   beforeAll(async () => {