+++ tests/unit/dispatchers/dispatchersService.test.ts
@@ -10,7 +10,7 @@ import { dispatcherData } from '../../__helpers__/dispatcher';
 import { createPagedResponse } from '../../__helpers__/utils';
 
 describe('Dispatchers Service', () => {
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
   const dispatcherId = dispatcherData.id;
   const dispatchersResponse = createPagedResponse(dispatcherData);