+++ tests/unit/events/healthService.test.ts
@@ -3,7 +3,7 @@ import { AUTH_BASE, EVENTS_BASE } from '../../../src/constants';
 import { Client, createClient, ParamsOauth2 } from '../../../src/index';
 
 describe('Events - Health Service', () => {
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
 
   let sdk: Client<ParamsOauth2>;