+++ tests/unit/localizations/languagesService.test.ts
@@ -4,7 +4,7 @@ import { Client, createClient, ParamsOauth2 } from '../../../src/index';
 import { createPagedResponse } from '../../__helpers__/utils';
 
 describe('Languages Service', () => {
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
 
   let sdk: Client<ParamsOauth2>;