+++ tests/unit/payments/appStoreService.test.ts
@@ -12,7 +12,7 @@ import {
 import { createPagedResponse } from '../../__helpers__/utils';
 
 describe('App Store Service', () => {
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
 
   let sdk: Client<ParamsOauth2>;