+++ tests/unit/payments/playStoreHistoryService.test.ts
@@ -9,7 +9,7 @@ import {
 import { createPagedResponse } from '../../__helpers__/utils';
 
 describe('Play Store History Service', () => {
-  const host = 'https://api.xxx.fibricheck.com';
+  const host = 'https://api.xxx.extrahorizon.io';
 
   let sdk: Client<ParamsOauth2>;